services/bot/requirements.txt
discord.py
requests
redis
emoji